requirements.txt
fastapi
uvicorn
pandas
chromadb
google-genai
openpyxl
pydantic
python-multipart
